added access control, session messages, fixed bugs

+++ LyXSc/app/Http/Controllers/Backend/ComicsManagement/ChaptersController.php
@@ -9,6 +9,7 @@ use App\Models\Page;
 use App\Models\Volume;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
@@ -20,6 +21,7 @@ class ChaptersController extends Controller
     use AuthorizesRequests;
     public function uploadChapter(Chapter $chapter, Request $request)
     {
+       // $this->authorize('handle comic management',    Auth::user());
         $valcomic =  $this->validate($request, [
             'file' => 'mimes:jpg,png,jpeg,gif',
         ]);
@@ -67,6 +69,7 @@ class ChaptersController extends Controller
 
     public function storeChapter(Volume $volume, Request $request)
     {
+        $this->authorize('handle comic management',    Auth::user());
         $chapterdata =  $this->validate($request, [
             'name' => 'required|min:1|max:255|string',
             //  'upload_date' => 'string',
@@ -81,13 +84,13 @@ class ChaptersController extends Controller
                 'isLocked' => $chapterdata["isLocked"],
                 'volume_id' => $volume->id
             ]);
-          return   Redirect::route('comics_management.volume.view',   $volume->id);
+          return   Redirect::route('comics_management.volume.view',   $volume->id)->with('message', 'Chapter Successfully Created.');
         }
         catch(\Exception $e){
                // $this->emit('Danger_alert',   'Something went wrong~ Try again~');
              //   $this->error = 'Something went wrong! try again~';
                 //  return dd($e);
-                Redirect::route('comics_management.volume.view',   $volume->id);
+                Redirect::route('comics_management.volume.view',   $volume->id)->with('error', 'Chapter Not Created.');
                 throw $e;
             }
         //$this->reset();
@@ -95,7 +98,7 @@ class ChaptersController extends Controller
 
     public function createVolumeStore(Comic $comic ,Request $request )
     {
-
+        $this->authorize('handle comic management',    Auth::user());
        //  dd($request);
         $volume_c =  $this->validate($request, [
             'name' => 'required|min:1|max:255|string',
@@ -113,14 +116,14 @@ class ChaptersController extends Controller
             ]);
            // $volume->save();
            //redirect('welcome');
-           return Redirect::route('comics_management.comics.view',   $comic->id);
+           return Redirect::route('comics_management.comics.view',   $comic->id)->with('message', 'Volume Successfully Created.');
 
         }
         catch(\Exception $e){
            //     $this->emit('Danger_alert',   'Something went wrong~ Try again~');
              //   $this->error = 'Something went wrong! try again~';
               //dd($e);
-              Redirect::route('comics_management.comics.view',   $comic->id);
+              Redirect::route('comics_management.comics.view',   $comic->id)->with('error', 'Volume Not Created.');
              throw $e;
             }
         //$this->reset();
@@ -128,6 +131,7 @@ class ChaptersController extends Controller
     }
     public function viewVolume(Volume $volume)
     {
+        $this->authorize('view comic management',    Auth::user());
         return Inertia::render('Backend/ComicsManagement/Volumes/Actions/viewVolume', [
             'volume' => [
                 'id' => $volume->id,
@@ -139,7 +143,7 @@ class ChaptersController extends Controller
                 'comic_id' => $volume->comic_id
             ],
 
-            'chapters' =>   Chapter::Where('volume_id', $volume->id)->get()->map(function ($chapter) {
+            'chapters' =>   Chapter::Where('volume_id', $volume->id)->orderBy('name', 'desc')->get()->map(function ($chapter) {
                 return [
                 'id' => $chapter->id,
                 'name' => $chapter->name,
@@ -156,30 +160,32 @@ class ChaptersController extends Controller
     }
     public function deleteChapter(Chapter $chapter)
     {
+        $this->authorize('handle comic management',    Auth::user());
        // dd($chapter);
        try{
         $chapter->delete();
-         return redirect()->back();
+         return redirect()->back()->with('error', 'Deleted Successfully.');
        }catch( \Exception $e){
             throw $e;
        }
     }
     public function deleteVolume(Volume $volume)
     {
+        $this->authorize('handle comic management',    Auth::user());
        try{
         $volume->delete();
-        return redirect()->back();
+        return redirect()->back()->with('error', 'Deleted Successfully.');
        }catch( \Exception $e){
             throw $e;
        }
     }
     public function deletePage(Page $page)
     {
-
+        $this->authorize('handle comic management',    Auth::user());
       //  dd($page);
        try{
         $page->delete();
-        return redirect()->back();
+        return redirect()->back()->with('error', 'Deleted Successfully.');
        }catch( \Exception $e){
             throw $e;
        }
@@ -187,7 +193,7 @@ class ChaptersController extends Controller
 
     public function viewChapter(Chapter $chapter)
     {
-
+        $this->authorize('view comic management',    Auth::user());
         return Inertia::render('Backend/ComicsManagement/Chapters/Actions/viewChapter', [
             'chapter' => [
                 'id' => $chapter->id,
@@ -200,10 +206,12 @@ class ChaptersController extends Controller
                 'volume_number' => $chapter->volume->number
             ],
 
-            'pages' =>   Page::Where('chapter_id', $chapter->id)->get()->map(function ($page) {
+            'pages' =>   Page::Where('chapter_id', $chapter->id)->orderBy('fileName', 'desc')->get()->map(function ($page) {
                 return [
                 'id' => $page->id,
-                'viewUrl' => 'comics_management.chapter.view',
+                'fileName' => $page->fileName,
+                'fileSize' => $page->fileSize,
+                'viewUrl' => route('comics_management.page.view', $page),
                 'createdAt' => $page->created_at,
                 'updatedAt' => $page->updated_at,
                 ];
@@ -211,6 +219,11 @@ class ChaptersController extends Controller
             'crftoken' => csrf_token()
 
         ]);
+    }
+
+    public function viewPage(Page $page)
+    {
+        return redirect($page->getFirstMediaUrl('page'));
     }
 
 }
